feat(#84): allow build on amd64 machines
